HELMET_BACKEND/requirements.txt
dotenv
fastapi
mysql-connector-python
pydantic
python-dotenv
asyncio
uvicorn[standard]
bcrypt
websockets
tidb_vector

# Core frameworks
fastapi
fastmcp
uvicorn

# LangChain ecosystem
langchain
langchain-anthropic
langchain-core
langchain-xai
langchain-community
langchain-mcp-adapters
langgraph

# Data validation and modeling
pydantic

# Environment and configuration
python-dotenv

# Cloud services
google-cloud-aiplatform
googlemaps
vertexai

# Database
mysql-connector-python
sqlalchemy
tidb-vector

# Date/time handling
pytz

# Utilities
pathlib